refactor: rename props variables and fix selection issue
Because the rendering was made by a component instead of being directly
inside the `edit.js`, the block was hard to select inside the editor.
By incorporating the function inside the edit file, the issue is gone.
So, I move some components files to obtain a more consistent structure.
I also rename the variable by declaring them on top of the functions to
keep consistency between the different appeals from `props`.